Move `impl TryFrom<StarkFelt> for EthAddress` next to struct definition.
